requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.6.0
Django==3.1.6
django-allauth==0.44.0
django-filter==2.4.0
djangorestframework==3.12.2
idna==2.10
importlib-metadata==3.4.0
Markdown==3.3.3
oauthlib==3.1.0
Pillow==8.1.0
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
zipp==3.4.0
